# src/main/resources/swagger/user-credit-sys.yaml
swagger: "2.0"
info:
  version: '1.0'
  title: User Management Business App
  description: This App exposes API to manage user like credit check
schemes:
  - http
basePath: /sys/user-credit/
paths:
  /v1/check-credit:
    x-rest-classname: UserCreditCheckV1
    post:
      tags:
      - User Credit
      summary: API to Check the Credit Score of the User
      operationId: creditCheck
      description: Check Credit Score of the user 
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
        - in: header
          name: MWMD-requestTimestamp
          description: >-
            Indicates the time at which request was created.
            Date format: ISO-8601 format (yyyy-MM-dd'T'HH:mm:ss'Z' or yyyy-MM-dd'T'HH:mm:ss±hh:mm)
          required: true
          type: string
        - in: header
          name: MWMD-activityName
          description: >-
            Identifies the name of the business activity for the message being
            exchanged.
            For MYUPC channel, this should be provided as "CustomerCare"
          required: true
          type: string
        - in: header
          name: MWMD-requestID
          description: >-
            RequestID is a unique identifier that is unique for a single
            request/response message and will be used to map log and error
            message to consumer’s request.
          required: true
          type: string
        - in: body
          name: userCreditCheckRequest
          description: >-
            Represents  Resource entity
          required: true
          schema:
            $ref: '#/definitions/UserCreditRequest'    
      responses:
        202:
          description: Accepted
        400:
          description: Bad Request
        500:
          description: Internal Server Error
        551:
          description: Unable to reach Downstream system
        552:
          description: External System Connection error
definitions:
  APIResponseView:
    properties:
      statusCode:
        type: string
        description: >-
          Indicates the status of the API response. Value of the status code
          indicates if it was a success or failure
          example: '2000000'
      statusDescription:
        type: string
        description: >-
            Defines the status of the API response. Status Description provides
            more information on the status
        example: SUCCESS
      severity:
        type: string
        description: >-
            Defines the severity of the API response in case of error response. 
            Possible values can be INFORMATION, WARNING, ERROR, CRITICAL
        example: INFORMATION
  UserCreditRequest:
    description: 'User Credit Request Object'
    properties:
      id:
        type: string
        description: User Id
      name:
        type: string
        description: User Name
  UserCreditResponse:
    description: 'User Credit Response Object'
    properties:
      id:
        type: string
        description: User Id
      name:
        type: string
        description: User Name 
      score:
        type: string
        description: Credit Score of the user             